# Improve interface performance using resolve_type