--- a/strawberry/object_type.py
+++ b/strawberry/object_type.py
@@ -139,6 +139,7 @@ def _process_type(
     interfaces = _get_interfaces(cls)
     fields = _get_fields(cls)
     is_type_of = getattr(cls, "is_type_of", None)
+    resolve_type = getattr(cls, "resolve_type", None)
 
     cls.__strawberry_definition__ = StrawberryObjectDefinition(
         name=name,
@@ -151,6 +152,7 @@ def _process_type(
         extend=extend,
         _fields=fields,
         is_type_of=is_type_of,
+        resolve_type=resolve_type,
     )
     # TODO: remove when deprecating _type_definition
     DeprecatedDescriptor(