--- a/strawberry/types/types.py
+++ b/strawberry/types/types.py
@@ -51,6 +51,9 @@ class StrawberryObjectDefinition(StrawberryType):
     extend: bool
     directives: Optional[Sequence[object]]
     is_type_of: Optional[Callable[[Any, GraphQLResolveInfo], bool]]
+    resolve_type: Optional[
+        Callable[[Any, GraphQLResolveInfo, GraphQLAbstractType], str]
+    ]
 
     _fields: List[StrawberryField]
 
@@ -97,6 +100,7 @@ class StrawberryObjectDefinition(StrawberryType):
             description=self.description,
             extend=self.extend,
             is_type_of=self.is_type_of,
+            resolve_type=self.resolve_type,
             _fields=fields,
             concrete_of=self,
             type_var_map=type_var_map,