This commit improves the performance of type resolution when returning interface types from a resolver by defining [`resolve_type`](https://github.com/strawberry-graphql/strawberry/pull/1949) that is called once per item. Without the existence of `resolve_type` the `is_type_of` method on each interface implementation is called until a `True` value is returned. See performance improvements below:

![resolve_performance_dracula](https://user-images.githubusercontent.com/29230371/172447974-612a371d-882c-4257-b6fe-cf9d873af46f.svg#gh-dark-mode-only)
![resolve_performance_dufte](https://user-images.githubusercontent.com/29230371/172447993-0ff1ee76-6a81-4860-a5a2-cd9bfcc3e551.svg#gh-light-mode-only)

<details>
  <summary>Benchmark Code</summary>

  ```python

  from collections import defaultdict
  from functools import partial
  from timeit import timeit
  from typing import DefaultDict, List, Sequence, Type, TypeVar
  
  import matplotx
  import numpy as np
  from matplotlib import pyplot as plt
  from perfplot._main import PerfplotData
  
  import strawberry
  from strawberry.scalars import ID
  
  
  T = TypeVar("T")
  
  
  @strawberry.interface
  class Item:
  
      id: ID
  
  
  @strawberry.interface
  class ResolvedItem:
  
      id: ID
  
      def resolve_type(self, *args):
          type_definition = self._type_definition  # type: ignore
          if isinstance(self, type_definition.origin):
              return type_definition.name
  
  
  def create_schema(n_items: int, n_types: int, item_type: Type) -> strawberry.Schema:
      CONCRETE_TYPES = []
      for i in range(n_types):
          CONCRETE_TYPES.append(strawberry.type(type(f"Item{i}", (item_type,), {})))
  
      items_list = [CONCRETE_TYPES[i % n_types](id=i) for i in range(n_items)]
  
      @strawberry.type
      class Query:
  
          items: List[item_type]  # type: ignore
  
      class Schema(strawberry.Schema):
          def execute_sync(self, *args, **kwargs):
              kwargs.update(root_value=Query(items=items_list))
              return super().execute_sync(*args, **kwargs)
  
      return Schema(query=Query, types=CONCRETE_TYPES)
  
  
  get_resolve_type_schema = partial(create_schema, item_type=ResolvedItem)
  get_is_type_of_schema = partial(create_schema, item_type=Item)
  
  query = "query { items { id } }"
  
  
  def benchmark(n_range: Sequence[int], n_repeats: int = 3):
      kernels = {
          "is_type_of (n_types = 2)": partial(get_is_type_of_schema, n_types=2),
          "is_type_of (n_types = 16)": partial(get_is_type_of_schema, n_types=16),
          "is_type_of (n_types = 32)": partial(get_is_type_of_schema, n_types=32),
          "resolve_type (n_types = 2)": partial(get_resolve_type_schema, n_types=2),
          "resolve_type (n_types = 16)": partial(get_resolve_type_schema, n_types=16),
          "resolve_type (n_types = 32)": partial(get_resolve_type_schema, n_types=32),
      }
      results: DefaultDict[str, List[float]] = defaultdict(list)
      for label, schema_fn in kernels.items():
          timings = results[label]
          for n_items in n_range:
              schema = schema_fn(n_items=n_items)
              repeats = []
              for _ in range(n_repeats):
                  repeats.append(timeit(lambda: schema.execute_sync(query), number=1))
              timings.append(min(repeats))
      timings = np.vstack([np.array(r) for r in results.values()])
      return PerfplotData(
          n_range=n_range,
          labels=tuple(kernels.keys()),
          xlabel="$n_{items}$",
          timings_s=timings,
          flop=None,
          title=None,
      )
  
  
  results = benchmark(n_range=[2 ** k for k in range(0, 13, 4)], n_repeats=10)
  
  for style in ("dufte", "dracula"):
      plt.style.use(getattr(matplotx.styles, style))
      results.save(
          f"resolve_performance_{style}.svg",
          transparent=False,
          bbox_inches="tight",
          relative_to=0,
      )
  
  ```
</details>

## Issues Fixed or Closed by This PR

- Closes #2320

## Types of Changes

<!--- What types of changes does your pull request introduce? Put an `x` in all the boxes that apply. -->
- [ ] Core
- [ ] Bugfix
- [x] New feature
- [x] Enhancement/optimization
- [ ] Documentation

# Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the CONTRIBUTING document.
- [x] I have added tests to cover my changes.
- [x] I have tested the changes and verified that they work and don't break anything (as well as I can manage).