@skilkis 2022-06-07
@patrick91 @bellini666 @BryceBeagle another area we can gain a tiny bit of performance. This is low priority, but I wanted to make this draft PR so we keep it in mind 😊