@skilkis 2023-07-13
Hey @iamkhav, thanks for the comment this PR has gone stale for a while. Implementation wise I'm done, I'll quickly rebase and push my latest changes. I see that @patrick91 recently implemented [`codspeed`](https://codspeed.io/) on this repo so that's very exciting. I'll add a simple benchmark with this PR. Would you be interested in contributing a benchmark as well? 🚀