@skilkis 2023-07-13
**Before**:
```shell
------------------------------------------------------------------------------------------ benchmark: 4 tests ------------------------------------------------------------------------------------------
Name (time in ms)                         Min                 Max                Mean            StdDev              Median               IQR            Outliers      OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_interface_performance[1]         18.0224 (1.0)       50.3003 (1.77)      20.0878 (1.0)      5.1488 (5.14)      18.9523 (1.0)      0.7415 (1.39)          2;4  49.7814 (1.0)          41           1
test_interface_performance[16]        22.0060 (1.22)      28.4240 (1.0)       22.8433 (1.14)     1.0016 (1.0)       22.6946 (1.20)     0.5321 (1.0)           3;3  43.7764 (0.88)         40           1
test_interface_performance[256]       69.1364 (3.84)      76.1349 (2.68)      72.2835 (3.60)     2.7917 (2.79)      72.1791 (3.81)     5.3980 (10.14)         4;0  13.8344 (0.28)          8           1
test_interface_performance[4096]     219.6461 (12.19)    231.3732 (8.14)     226.4553 (11.27)    4.4473 (4.44)     227.2840 (11.99)    5.8001 (10.90)         2;0   4.4159 (0.09)          5           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

**After**:
```shell
---------------------------------------------------------------------------------------- benchmark: 4 tests ----------------------------------------------------------------------------------------
Name (time in ms)                        Min                Max               Mean            StdDev             Median               IQR            Outliers      OPS            Rounds  Iterations
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_interface_performance[1]        14.3921 (1.0)      46.2064 (2.79)     16.3355 (1.05)     4.2380 (10.94)    15.5533 (1.0)      0.6572 (1.46)          2;5  61.2163 (0.96)         57           1
test_interface_performance[16]       14.8669 (1.03)     16.5732 (1.0)      15.6205 (1.0)      0.3874 (1.0)      15.6228 (1.00)     0.5508 (1.22)         19;0  64.0185 (1.0)          61           1
test_interface_performance[256]      15.8977 (1.10)     24.4618 (1.48)     17.5442 (1.12)     1.2834 (3.31)     17.2590 (1.11)     0.8033 (1.78)          7;4  56.9989 (0.89)         53           1
test_interface_performance[4096]     18.7340 (1.30)     21.2899 (1.28)     19.5104 (1.25)     0.4120 (1.06)     19.4837 (1.25)     0.4509 (1.0)          14;1  51.2548 (0.80)         51           1
----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```

This can be verified by commenting out the `resolve_type=_get_resolve_type()` argument in `GraphQLCoreConverter.from_interface`