@iamkhav 2023-07-14
Hey @skilkis, thanks so much for the quick response!
As requested, here are the benchmark results obtained from running the tests on my machine.

**Before:**
```bash
------------------------------------------------------------------------------------------- benchmark: 4 tests ------------------------------------------------------------------------------------------
Name (time in ms)                         Min                 Max                Mean            StdDev              Median               IQR            Outliers       OPS            Rounds  Iterations
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_interface_performance[1]          6.6333 (1.0)        8.3878 (1.0)        6.9347 (1.0)      0.2959 (1.0)        6.8261 (1.0)      0.3523 (1.0)          19;5  144.2030 (1.0)         123           1
test_interface_performance[16]         8.4697 (1.28)      10.3100 (1.23)       8.8984 (1.28)     0.4019 (1.36)       8.7234 (1.28)     0.6608 (1.88)         22;1  112.3792 (0.78)         76           1
test_interface_performance[256]       32.9080 (4.96)      35.2016 (4.20)      33.6217 (4.85)     0.7459 (2.52)      33.6041 (4.92)     0.8256 (2.34)          1;1   29.7427 (0.21)          9           1
test_interface_performance[4096]     111.7712 (16.85)    115.1568 (13.73)    113.4827 (16.36)    1.4652 (4.95)     113.6300 (16.65)    2.6300 (7.46)          2;0    8.8119 (0.06)          5           1
---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```
**After:**
```bash
--------------------------------------------------------------------------------------- benchmark: 4 tests ---------------------------------------------------------------------------------------
Name (time in ms)                       Min                Max              Mean            StdDev            Median               IQR            Outliers       OPS            Rounds  Iterations
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
test_interface_performance[1]        5.1455 (1.0)       6.2182 (1.0)      5.5483 (1.0)      0.1878 (1.0)      5.5265 (1.0)      0.2634 (1.0)          48;2  180.2352 (1.0)         157           1
test_interface_performance[16]       5.3090 (1.03)      6.9405 (1.12)     5.6823 (1.02)     0.2224 (1.18)     5.6597 (1.02)     0.3209 (1.22)         51;1  175.9864 (0.98)        164           1
test_interface_performance[256]      5.4854 (1.07)      7.0765 (1.14)     6.0096 (1.08)     0.3587 (1.91)     5.9452 (1.08)     0.5926 (2.25)         59;0  166.3992 (0.92)        164           1
test_interface_performance[4096]     6.1811 (1.20)     99.4073 (15.99)    7.6783 (1.38)     8.2804 (44.09)    6.8460 (1.24)     0.5935 (2.25)          1;1  130.2373 (0.72)        125           1
--------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
```